requirements.txt
# Core dependencies
requests>=2.28.0
python-dateutil>=2.8.2

# Google Calendar API
google-api-python-client>=2.70.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0

# Optional dependencies for enhanced functionality
# Uncomment as needed
# pandas>=1.5.0  # For data analysis
# matplotlib>=3.6.0  # For visualization
# schedule>=1.1.0  # For scheduling recurring tasks
